Rename HashableDict file
