+++ src/graph_notebook/gremlin/client_provider/graphsonV3d0_MapType_objectify_patch.py
@@ -4,7 +4,7 @@ SPDX-License-Identifier: Apache-2.0
 """
 
 from gremlin_python.structure.io.graphsonV3d0 import MapType
-from graph_notebook.gremlin.client_provider.utils_tinkerpop import HashableDict
+from graph_notebook.gremlin.client_provider.hashable_dict_patch import HashableDict
 
 # Original code from Tinkerpop 3.4.1
 #
