+++ src/graph_notebook/gremlin/client_provider/hashable_dict_patch.py
